chore: rename moduleUtils.js to module-utils.js
