+++ lib/debugger-wrapper.js
@@ -3,7 +3,7 @@ const debug = require('debug')('snyk:nodejs:inspector');
 const inspector = require('inspector');
 
 const vulnMgmt = require('./vuln-mgmt');
-const moduleUtils = require('./moduleUtils');
+const moduleUtils = require('./module-utils');
 const transmitter = require('./transmitter');
 
 var session;
